ai-services/requirements.txt
# Core Framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# AI/ML Models
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
timm>=0.9.0

# Image Processing
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0

# OCR
paddlepaddle>=2.5.0
paddleocr>=2.7.0
pytesseract>=0.3.10

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Color Extraction
colorthief>=0.2.1
scikit-learn>=1.3.0

# Data Processing
pandas>=2.1.0
